requirements.txt
Flask==3.1.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.27.1
psycopg2-binary==2.9.10
transformers==4.46.3
PyPDF2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.4
python-dotenv==1.0.1
pytest==7.3.1
pytest-flask==1.2.0
torch==2.0.1 
torchvision==0.15.2
torchaudio==2.0.2
sentencepiece==0.1.96